Gather all resolved calls, filter for concrete ones
This fixes an issue in test-generator where a call to `toString` is dispatched
to the `toString` method from `java.lang.Object` is used instead of the
overridden method from `ArrayList`.

Cite from the issue about the current implementation:

> It tracks a `visited` set preventing it from listing callees twice when
> multiple inheritance is in play (e.g. Java interfaces). Unfortunately this
> malfunctions when we visit a type twice, once via its interfaces and once via
> its concrete subclass which provides a definition

This fix removes the `visited` set and therefore traverses the whole class
hierarchy (in the case of java.lang.Object, else a sub-tree) to find appropriate
implementation methods.